requirements.txt
streamlit==1.28.1
pandas==2.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
streamlit-autorefresh==0.0.1
Pillow==10.0.0
plotly==5.15.0